refactored API functions, added googleplaces searchbar
Earlier functions for fetching API data  were borrowed from a previous version of a different weather app I started earlier, but they were overly
complicated and I refactored them for simplicity. I'm using a geocoding API to get specific location data to send to the one call
API from a search text, but sometimes the geocoding API returns a JSON object in which the first sub-object is not correct, so rather than
returning the full JSON object, I just return the object I am (or the user is)  looking for.(This is all in forecast.js.) Then I can handle this data in
index.js. Next up is to format the DOM with the data.