Update unix time function to return either time or day
Move unixTime to index.js
Rename for simplicity and convert it to a factory function that
can either return the day or time from the unix string. This will make
it easier to reuse for multiple purposes.
Update any previous uses of the unixTime() function to work with rename
and retooling.