Change geocoding api to google's
Update the functions that rely on the geocoding api to work with the google
geocoding api. The previous api only worked with http, and had issues in firefox,
so this commit will update the app to work with the google api, which works more
consistently across browsers, and is more secure.